Add files via upload
Please take note: This website contains some files that have identical names. I had to alter their names since this platform couldn't allow me to upload the files in their folders. Take a look at the file labeled structure and insert the files in their respective folders. For example, files index1 (rename it to index), dashboard, reports, found_items and logout should be in a folder labeled admin. Also rename hearder1 to header.  As for files logo and logo1, leave them where they are. As for the rest, follow the instructions above with the aid of the file labeled structure. Zimlugrecovery123! is the admin password to the admin dashboard